--- app/component/CookieBasedClient.ts
@@ -1,11 +1,0 @@
-import { type Schema } from '@/amplify/data/resource';
-import outputs from '@/amplify_outputs.json';
-import { generateServerClientUsingCookies } from '@aws-amplify/adapter-nextjs/data';
-import { cookies } from 'next/headers';
-
-// https://docs.amplify.aws/nextjs/build-a-backend/data/connect-from-server-runtime/nextjs-server-runtime/
-export const cookieBasedClient = generateServerClientUsingCookies<Schema>({
-  config: outputs,
-  cookies,
-  authMode: "userPool"
-});
